Commit Process

+++ dataset.py
@@ -1046,7 +1046,8 @@ class TCIA(Dataset):
         slice_name = self.sample_list[idx]
         count = slice_name.split('_')[1]
         x = int(slice_name.split('_')[2].split('slice')[1][0:3])
-        label = torch.tensor(round((x / self.index[count]) * 60.0))
+        label = torch.tensor(int(round((x / self.index[count]) * 60.0)))
+
         data_path = os.path.join(self.data_dir, slice_name)
         data = np.load(data_path)
         image = data['image'], data['label']
